dashboard/api/pyproject.toml
[project]
name = "mem0-dashboard-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "python-multipart>=0.0.9",
    "mem0ai>=0.1.92",
    "qdrant-client>=1.12.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
graph = ["neo4j>=5.0.0", "langchain-neo4j>=0.1.0", "rank-bm25>=0.2.0"]
test = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
